agent/requirements.txt
langchain-groq==0.0.10
langchain-core==0.1.33
langgraph==0.0.27
pinecone-client==3.0.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.3.0
typing-extensions==4.9.0
email-validator==2.1.0 